charinfo: move parsing to separate file
